Change RotaryEncoder code to boards.py
